Added error messages , login functionality and routes

+++ src/kit/api/User.js
@@ -10,9 +10,10 @@ export const register = async (details) => {
         body: JSON.stringify(details)
     }
 
-    await fetch('http://localhost:5000/user/register', user)
-        .then(res => {return res})
-        .catch(error => console.log(error))
+    return await fetch('http://localhost:5000/user/register', user)
+                 .then(res => res.text())
+                 .then(data => sessionStorage.setItem('registration', data))
+                 .catch(error => console.log(error))
 
 }
 export const login = async (details) => {
@@ -26,9 +27,15 @@ export const login = async (details) => {
     }
 
     await fetch('http://localhost:5000/user/login', post)
+        .then(data => data.json())
         .then(res => {
+            if (res === "Username doesn't exist!" || res === 'Wrong password!') {
+                sessionStorage.setItem('error', `${res}`)
+                return
+            }
+
             sessionStorage.setItem('token', `Bearer ${res}`)
-            return 'Login successful!'
+         
         })
         .catch(error => console.log(error))
